+++ CHANGELOG.md
@@ -16,6 +16,7 @@ HumHub Changelog
 - Fix #6795: Fix searching of disabled users in Administration
 - Fix #6834: Fix notification switcher in wall stream entry context menu
 - Fix #6836: Fix check permission of changing content to public from context menu
+- Fix #6837: Fix saving of public content
 
 1.15.2 (December 19, 2023)
 --------------------------